+++ tests/test_tpa_pipeline.py
@@ -148,6 +148,7 @@ class TestTpaPipeline(unittest.TestCase):
         kwargs = {'new_association': new_association}
         backend = self.get_mocked_sso_backend()
         backend.name = backend_name
+        backend.strategy.session_get.return_value = 'not-an-enrollment-url'
         self.user = UserFactory(is_active=True)
         multiple_enterprises_feature.return_value = multiple_enterprise_switch
         enterprise_customer = EnterpriseCustomerFactory(
@@ -190,6 +191,7 @@ class TestTpaPipeline(unittest.TestCase):
         kwargs = {'new_association': new_association}
         backend = self.get_mocked_sso_backend()
         backend.name = backend_name
+        backend.strategy.session_get.return_value = 'not-an-enrollment-url'
         self.user = UserFactory(is_active=True)
         multiple_enterprises_feature.return_value = True
         enterprise_customer = EnterpriseCustomerFactory(
@@ -206,6 +208,71 @@ class TestTpaPipeline(unittest.TestCase):
                 fake_get_ec.return_value = None
                 handle_enterprise_logistration(backend, self.user, **kwargs)
                 ent_page_redirect.assert_not_called()
+
+    @ddt.data(
+        (True, False, True, 'facebook'),
+        (False, False, True, 'facebook'),
+        (True, True, False, 'facebook'),
+        (False, True, False, 'facebook'),
+        (True, True, False, 'facebook'),
+        (False, True, False, 'facebook'),
+        (True, False, True, 'facebook'),
+        (False, False, True, 'facebook'),
+        (True, True, True, 'google-oauth2'),
+        (False, True, True, 'google-oauth2'),
+        (True, False, False, 'google-oauth2'),
+        (False, False, False, 'google-oauth2'),
+        (True, False, True, 'google-oauth2'),
+        (False, False, True, 'google-oauth2'),
+        (True, True, False, 'google-oauth2'),
+        (False, True, False, 'google-oauth2'),
+    )
+    @ddt.unpack
+    @mock.patch('enterprise.tpa_pipeline.is_multiple_user_enterprises_feature_enabled')
+    def test_bypass_enterprise_selection_page_for_enrollment_url_login(self,
+                                                                       using_enrollment_url,
+                                                                       new_association,
+                                                                       multiple_enterprise_switch,
+                                                                       backend_name,
+                                                                       multiple_enterprises_feature):
+        """
+        Test that enterprise selection page is bypassed if socialAuth user is part of multiple enterprises
+        and uses an enrollment url for login
+        """
+        kwargs = {'new_association': new_association}
+        backend = self.get_mocked_sso_backend()
+        backend.name = backend_name
+        if using_enrollment_url:
+            backend.strategy.session_get.return_value = '/enterprise/12e87171-fb0a/course/course-v1:Test/enroll'
+        else:
+            backend.strategy.session_get.return_value = 'not-an-enrollment-url'
+        self.user = UserFactory(is_active=True)
+        multiple_enterprises_feature.return_value = multiple_enterprise_switch
+        enterprise_customer = EnterpriseCustomerFactory(
+            enable_data_sharing_consent=False
+        )
+        enterprise_customer_old = EnterpriseCustomerFactory(
+            enable_data_sharing_consent=False
+        )
+        EnterpriseCustomerUser.objects.create(
+            enterprise_customer=enterprise_customer_old,
+            user_id=self.user.id,
+            active=False
+        )
+        EnterpriseCustomerUser.objects.create(
+            enterprise_customer=enterprise_customer,
+            user_id=self.user.id,
+            active=True
+        )
+        with mock.patch('enterprise.tpa_pipeline.get_enterprise_customer_for_running_pipeline') as fake_get_ec:
+            with mock.patch(
+                    'enterprise.tpa_pipeline.select_enterprise_page_as_redirect_url') as ent_page_redirect:  # pylint: disable=invalid-name
+                fake_get_ec.return_value = None
+                handle_enterprise_logistration(backend, self.user, **kwargs)
+                if new_association or not multiple_enterprise_switch or using_enrollment_url:
+                    ent_page_redirect.assert_not_called()
+                else:
+                    ent_page_redirect.called_once()
 
     def test_get_ec_for_pipeline(self):
         """